functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
google-generativeai>=0.8.0
anthropic>=0.40.0
